Extract BeefyGossipValidator
